simplifed SecLib signers
