+++ src/Namshi/JOSE/Signer/SecLib/RS384.php
@@ -4,10 +4,12 @@ namespace Namshi\JOSE\Signer\SecLib;
 
 class RS384 extends RSA
 {
-    public function __construct()
+    /**
+     * {@inheritdoc}
+     */
+    protected function configureEncryptionAlgorithm(\Crypt_RSA $encryptionAlgorithm)
     {
-        parent::__construct();
-        $this->encryptionAlgorithm->setHash('sha384');
-        $this->encryptionAlgorithm->setMGFHash('sha384');
+        $encryptionAlgorithm->setHash('sha384');
+        $encryptionAlgorithm->setMGFHash('sha384');
     }
 }
